+++ app/models/plan.rb
@@ -1,0 +1,4 @@
+class Plan < ApplicationRecord
+  self.implicit_order_column = "created_at"
+  has_many :questions
+end