+++ db/migrate/20201027160912_create_plan_table.rb
@@ -1,0 +1,8 @@
+class CreatePlanTable < ActiveRecord::Migration[6.0]
+  def change
+    create_table :plans, id: :uuid do |t|
+      t.string :category, null: false
+      t.timestamps
+    end
+  end
+end