+++ spec/factories/answer.rb
@@ -1,0 +1,7 @@
+FactoryBot.define do
+  factory :answer do
+    association :question, factory: :question
+
+    response { "Green" }
+  end
+end